fix: cyclic dependency between LocalizerUtil and SanitizationUtil (#14508)
